Align the start line of the prelude with start of the source file (#1512)
* align the start line of the prelude with start of the source file

This way the potential error messages originating from the prelude
won't need to be massaged to be legible. It will also help if one day
we want to _step into_ the prelude while debugging. One would
only (soft/hard-)link the prelude by side of the source and (hopefully!)
the debugger would pick it up.

This necessitated the movement of the OCaml comment inside of the
prelude, and as such it now became a Motoko comment...

I'd wager to say that this is a negligible parsing overhead.

* minimally simplify prelude

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>