DFX-Interface: Imports of _code_ of another canister
In meeting with @crusso and @matthewhammer about dynamic actor creation,
we converged on Claudio’s proposal that a clean, easy and desirable
way to expose dynamic actor creation in Motoko is to get some help from
`dfx`.

Big plus of this approach: You can, in one `dfx`, project, have a Motoko
and a Rust canister module, and the Motoko code can “import” the rust
“code” and then dynamically install instances of that code. So this
allows you write high-performance infinitely scaleable Rust backend
canisters together with a nice declarative Motoko canister that
orchestrates things. But thanks to Candid it also works nicely within
Motoko.

Another big approach is that is seems the fasted way towards clean,
hackfree dynamic actor creation.

The rough idea is that in Motoko you can write

    import Worker "class:worker";

and have that to be equivalent to `actor class Worker(params) body`
where `params` is imported (via a Candid-like file, `.params`) and the
`body` is also imported (by reading the actual `.wasm` file).

The impact on `dfx` is that it would do more of the stuff that it
currently does with the `.did` files: Figure out which canisters depend
on which, get build artifacts from `moc` (now: `.did`, then: also the
`.wasm` (which will be used the Motoko to install) and `.params` (so
that Motoko knows how to type the the imported class)) and making them
available to the using canister.

(The `.params` may also be useful for `dfx` to type-check the argument
to `canister canister install` for parametrized canisters.)

All filenames and other names up for bikeshedding.

This was written together by @crusso and @matthewhammer.