Remove old users’ guide (#1566)
as per https://dfinity.slack.com/archives/CPL67E7MX/p1591349237034900

I was unable to set up a similar setup for the new users guide using
antora, as it is far from trivial to “just run antora” on the
documentation “module” in this repository without jumping through hoops
like creating a temporary local git repository. (as
https://github.com/dfinity-lab/documentation/blob/master/antora.nix
does). This means that #1415 is still open (and more pressing)