Nix: Check that grammar.txt is up-to-date
generated files in repos are bad. Generated files in repos without a
check that they are up-to-date are even worse.

This adds a nix check that `grammar.txt` is up do date.

It also extracts the grammar producing code into a separate script (less
logic in Make)

I thought it would solve the problem of sed being different in Linux and
Darwin, by running CI on both, but CI runs of course `nix`’s `sed` on
both. Possible solution: Always run

    nix-shell --command "make -C src grammar"

or rewrite in `perl` ;-)