preparation: factor out code wrapping for declarations
This is part of the DWARF generation effort, as that will insert
instrumentation *around* current code (e.g. lexical region
boundaries). Since declarations are the source of new locals, it
is`compile_decs`, that is getting a type change in this commit.

Whereas formerly we'd return a piece of code that establishes the
binding(s), and which is inended to be prepended to the continuation
code, now we return a function, that when applied on the same will
itself prepend (and in future append) the necessary fragments.