IR: pass prelude through IR passes
this is prerequisite for adding convenience code into `prelude.ml` that
uses `async`/`await`, and a first step towards dynamic actor creation.

The change is actually a simplification: The prelude is now treated more
like an implicitly imported library, and is (until we have incremental
compilation) combined with the libraries and main program in
`desguar.ml`.