Nix: Check that grammar.txt is up-to-date (#1575)
generated files in repos are bad. Generated files in repos without a
check that they are up-to-date are even worse.

This adds a nix check that `grammar.txt` is up do date.

It also extracts the grammar producing code into a separate script (less
logic in Make)

This also solves the problem of sed being different in Linux and
Darwin, by making the `nix` version (inside `nix-shell`) canonical.

If the script can be made to work with OSX’s `sed`, that's great, but as
we can't test that properly, it is in a way optional.