preparation: factor out code wrapping for declarations (#1576)
This is part of the DWARF generation effort (#1163), as that will insert
instrumentation *around* current code (e.g. lexical region
boundaries). Since declarations are the source of new locals, it
is `compile_decs`, that is getting a type change in this commit.

Whereas formerly we'd return a piece of code that establishes the
binding(s), and which is intended to be prepended to the continuation
code, now we return a function, that when applied on the same will
itself prepend (and in future append) the necessary fragments.

(N.B.: this is somewhat reminiscent of HOAS, but has different mechanics.)

The convention is that code fragments that wrap other code have `W` in
their binding names. `unmodified` is provided as the trivial wrapper.

Includes a few cleanups and typofixes.