Warn when the read file has a different name from the requested file (#1623)
When a file is opened with name `foo.mo` on a case insensitive file system, and there is no `foo.mo` file present, but a `Foo.mo`, then the open will succeed, but on a machine with a case-sensitive file system the open will fail. (E.g. for files checked a repository.) We now issue a warning when the directory doesn't contain a file with the exact same spelling.

Fixes #1435.