Inject a source identifer into `moc` (#1631)
I have been preaching this for so long to other teams, I really better
clean up my own backyard: Make sure that you can run

    moc --version

and get a somewhat useful indication what version that is. This is
important for when random people use `moc` from random versions of SDK
or – later – build it themselves and then provide bug reports.

When building locally, with `.git` around, this now injects the git
revision:

    $ ./moc --version
    Motoko compiler (revision 9abae15e-dirty)

(the `-dirty` indicates that my work-tree is not clean).

When `.git` is _not_ available, as typically (and rightfully) when
building with nix, it looks if Nix’s `$out` variable is set, and
includes that:

    $ $(nix-build -A moc)/bin/moc --version
    Motoko compiler (revision 3yrjbw6g-aqppc05d-596yiq8k-bsznw5jw)

This is less useful for end-users, but it still allows us to (hopefully)
identify the version, in the worst case by running this command on all
revisions to find the right one:

    $ nix-store --query --outputs $(nix-instantiate -A moc-bin)
    /nix/store/3yrjbw6gaqppc05d596yiq8kbsznw5jw-moc-bin

Note that I injected dashes into the id so that it does not trip up 
`common.lib.standaloneRust` which otherwise would take it for an
unwanted dependency.

@basvandijk suggests in
https://github.com/dfinity-lab/sdk/pull/383#issuecomment-588515223 that one
may query Hydra’s database to map from id to revision.

Once we do versioned releases, we can of course extend this to inject
the version number there, and use `git describe --tag` to have
`v1.0-200-9abae15e` like descriptions.

I _hope_ this will not cause extra recompilations with `dune`, and will
work out-of-the box for everyone, but we’ll see.