Implement candid prims
I need the PR prims for dynamic creation of actor classes; the source
exposition is mainly just for testing.

It follows the example of `debug_show`. Obviously this kinda does not
scale, and we need a better interface here. But maybe in a separate PR.

(One idea: We already kinda overload `e1(e2)`, given that we have
different arrows. So how about an abstract type `MagicShowFun` that is
_not_ a subtype of some function, but can still be called via `e1(e2)`
syntax. This way, `debug_show` etc. can be come normal “values” that can
be let-bound and thus expose via the standard library.)

Will pull out some refactoring separately.