Tagged scalars: Just shift, don't rotate
Previously we had

    0bxxx0z: number 0x0zxxx (arithmetic rotation by 2)
    0x00000: false
    0x00001: true
    0x00011: pointer

but it’s much nicer to have

    0bxxxx0: number 0x0xxxx (arithmetic right-shift by 1)
    0x00000: false
    0x00001: true
    0x00011: pointer

This means that the MakeCompact code does not need to distinguish
between “compressed” and “right-0-padded” representation; corresponding
code that rotates disappears. And we can refer to `BitTagged` for more
things as well instead of hardcoding arithmetic stuff.

This also moves functions defined in `BoxedWord64` that have nothing
to do with boxing into a module `Word64`.