Implement the GC in Rust
This implements the current garbage collector in Rust. No changes were
made to the GC design -- it's just ports the one implemented in code
generator to Rust.

The goals are:

- Evaluate Rust for Motoko's RTS implementation
- Make the collector easier to read, understand, modify, and extend.

Current status: the code is complete in the sense that there aren't any
missing features/passes etc., but it's has bugs. I'm not sure how to
debug Wasm yet. There are also lots of TODOs in the code, mostly for
documentation.

Submitting a PR to get early feedback.