CI: Block on darwin builds again (#1715)
in 513cb10b2f16e05acec349bf7961f8c42bd259d1 we stoppped blocking on
darwin builds on CI, so that PRs merge faster, but it is causing
friction when a darwin-specific build failure slips through to master.

Maybe we have faster CI builders now? Let’s try to revert this and see
if waiting for darwin is annoying again or not.