[mo-doc] Reuses parsing logic from pipeline (#1722)
This way we get nice error reporting for parse errors instead of cryptic exceptions.

I'm not excited about all the "primed" versions of functions in `pipeline.ml` but I also don't see an easy way to get rid of them. Maybe we'll need to think about some of the abstraction boundaries drawn in `pipeline`.

Maybe we'll decide doc generation and the language server shouldn't be clients of that module in the first place, or we'll need more abstractions and data types (the tupling gets a little out of hand).