Test suite: Fix actor reference types
we had a test case that checks parsing of an actor id in an `actor "…"`
declaration, but it didn’t work – parsing happens dynamically.

This moves that test to `drun` to be able to catch the error.

This also reveals issues in the interpreter, and in `ic-ref-run`. See
output files…