nix-shell: Put our own binaries onto the PATH
TL;DR: In the nix shell, just run `moc` from anywhere to run the `moc` you just built.

When building something like our test suite within `nix-build`, all the programs that the test suite needs (e.g. `moc`) are nicely on the `$PATH`. This would make writing scripts pretty easy and straight-forward.

But in the `nix-shell` environment, these build artifacts are _not_ built and put on the PATH by nix, for good reasons: The developer wants to build them manually and iteratively, and maybe with different flags (e.g. so that they can test it while the compiler still prints warnings).

Previously, everywhere where we wanted to invoke `moc` we had rather complex logic that would, by default, look in `../src/moc`, and then override that logic by setting `MOC=moc` from nix-build. This worked ok, but:

 * The logic obscures meaning (see the changes to the `test/repl` scripts).
 * The logic is duplicated.
 * When you want to run `moc` directly on a test file, you’d have to say `../../src/moc -i empty.mo` or something.
 * If `src/moc` was not built you get an unhelpful error message or stuff breaks randomly.

With this change, we now have `moc` on the `$PATH` even in the nix shell environment. This means that all tests/script work the same, and can be simplified (no more `$MOC` variables!). The `moc` in the `$PATH` is actually a generic wrapper script that

 * Checks if `src/moc` has been built. If not, it aborts, and even tells the developer the command to use to build it. E.g.:

        moc: target src/moc not built yet.
        To build, run:
            make -C ../../src moc

 * Applies the same logic to things like the `mo-rts.wasm` file: It checks it is there, if not aborts and tells the user how to build it, and if it is there, sets the `MOC_RTS`, again mimicing the “real” environment.

The script does _not_ ensure that `moc` is up-to-date. I believe that this would be too intrusive and possibly slow; I should be able to edit the code while the test suite is running without some obscure background builds happening.

If this works out nicely we can maybe export this idea to `dfinity`, who are facing similar problems.

Because `bin/wrapper.sh` is only used inside the `nix-shell` it is not CI-tested, and we will rely on developer feedback to notice issues with it.