Implement candid prims (IR only)
I need the IDL prims for dynamic creation of actor classes eventually.

This has _no_ source exposition (difference to previous PR #1652).
Instead, to test this from a test file, the magic `MOC_UNLOCK_PRIM`
variable has to be set (introduced in #1751) and the `prim "serialize"`
syntax has to be used.

The IR type-checker would catch invalid use of the `prim "serialize"`
thing.