Tag bools on the heap
In
https://github.com/dfinity-lab/motoko/pull/1693#issuecomment-656045323
@ggreif argued that recognizing pointers via `p & 3 == 3` is one
instruction more expensive than just checking a single bit. And on that
PR we have a +2.4% gas consumption regression.

So in this PR, I try to use just the lowest bit as the sentinely. This
requires shifting booleans when we put them into `Vanialla`
representation. Let’s see what CI says…