Test nixpkgs bumps weekly, but do not commit them
previously, every monday the bot would update `nixpkgs`. This is useful
because we want to get early feedback if upgrading would break anything
(so that when we _do_ need to upgrade, we don’t then have to deal with
any such breakage). But it also adds churn to `master`, including
`nix-shell` having to re-download stuff, without any actual gain.

So after this PR we still get a PR every monday sunday with the bump.
But if CI passes, then mergify will simply close the PR, _without_
committing it to master. So if all is well, nothing really happens.

But should the upgrade fail for some reason, the PR will stay open, as a
reminder to investiage.