Backend: Move allocate-on-use logic into lib (#1759)
this is a pure refactoring that moves the alloc-on-use logic that we
already use for built in (RTS-internal) functions into a dedicated
`Lib` module. After this refactoring, #1752 should look nicer and easier
to review.

This also introduces `named_imports`, to no longer track keep track of
imported functions together with built-in functions.