Tagged scalars: Just shift, don't rotate (#1693)
Previously we had

    0bx₂x₃x₄0x₁: number 0bx₁x₁x₂x₃x₄ (arithmetic rotation by 2)
    0b00000: false
    0b00001: true
    0b00011: pointer

but it’s much nicer to have

    0bx₁x₂x₃x₄0: number 0bx₁x₁x₂x₃x₄ (arithmetic right-shift by 1)
    0b00000: false
    0b00001: true
    0b00011: pointer

This means that the MakeCompact code does not need to distinguish
between “compressed” and “right-0-padded” representation; corresponding
code that rotates disappears. And we can refer to `BitTagged` for more
things as well instead of hardcoding arithmetic stuff.

This also moves functions defined in `BoxedWord64` that have nothing
to do with boxing into a module `Word64`.