Embed mo-rts.wasm into moc binary (#1772)
prompted by a discussion with @kritzcreek about releasing `moc` binaries
separtely I thought about how we can hide `mo-rts.wasm` from our users
(both direct downloaders and `dfx`).

This change will embed `mo-rts.wasm` statically into `moc` when built
with nix (by generating a source file with a long string literal).

For local development and testing, that would be annoying, though, so in
that environment, `moc` loads the RTS at runtime from a file. It learns
the location of that file via the `MOC_RTS` environment variable, but
that is invisible to developers if they use the `bin/moc` wrapper, i.e.
just run `moc` as it is in the PATH.