for C-style enumeration variants do the dereferencing ourselves
Don't leave it to the debugger-side dwarf. Wasmtime conjures up
a strange `WebAssemblyRefWrapper<??>` for anything that has a pointer
(or reference) type. But it turns out that when I declare the argument
type to be the C-like enum, and get the hash by a DWARF expression
(argument plus 5 (unskew and hop over the heap tag), dereferenced)
the `DW_OP_deref` will take care of adding the memory base in the
wasmtime translation step.