Backend: Dedicated SR.Bool
previously, we would use `box_i32` and `unbox_i32` when putting `bool`s
on the heap. That is correct, but wasteful: These functions check if the
value is small before shifting, but a boolean is always 0 or 1, so
always small.

This introduces a dedicated StackRep.UnboxedBool, which converts to and
from Vanilla more efficiently (simply shift).