.did parser: Handle byte sequences correctly
Previously, `"\ff"` byte sequences were mishandled. This shows up when
parsing Candid test assertions, but also when using these byte sequences
to write (valid) utf8 files.

I simply copied the code that we use in the Motoko lexer from
https://github.com/dfinity-lab/motoko/commit/f76d9a5dc9dde9d76e57d522fa7c3974ccee1c97

It is now the parser’s responsibility to check for bad utf8; I was not
sure how best to implement errors in Ocaml code in the parser. Or should
this check be moved to the candid type checker (like it is for Motoko)?

The ability to parse raw bytes is not used at the moment, but will be
with #1830.