.did parser: Handle byte sequences correctly (#1831)
Previously, `"\ff"` byte sequences were mishandled. This shows up when
parsing Candid test assertions, but also when using these byte sequences
to write (valid) utf8 files.

I simply copied the code that we use in the Motoko lexer from
https://github.com/dfinity-lab/motoko/commit/f76d9a5dc9dde9d76e57d522fa7c3974ccee1c97
and  https://github.com/dfinity-lab/motoko/pull/1835

It is now the parser’s responsibility to check for bad utf8.

The ability to parse raw blobs is not used at the moment, but will be
with #1830.