Bump niv to allow overriding of dependencies.
this allows us to use, say,

    NIV_OVERRIDE_dfinity=~/dfinity/dfinity nix-build
    NIV_OVERRIDE_motoko_base=~/dfinity/motoko-base nix-build
    etc.

to override the source specified in `nix/sources.json`. We had that
feature already for dfinity in a local hack, but since
https://github.com/nmattia/niv/pull/260 it’s official and thus works for
all sources mentioned in `nix/sources.json`.

This is super useful for bisecting a bug in `dfinity`, or testing things
with a local check-out of `motoko-base` etc.