Check Rust formatting in CI
This is not an ideal solution: we use stable rustfmt to check formatting
of files built with nightly compiler, which I think can fail if we use
new syntax available in nightly but not in stable yet.

Mozilla Rust overlay[1] doesn't have a derivation for nightly rustfmt so
this is the only way I could make this work without implementing my own
derivations.

[1]: https://github.com/mozilla/nixpkgs-mozilla