Class Imports; continued (#1845)
Implements support for importing Motoko compiled actor classes as libraries, supporting multi-canister applications such as DHTs (e.g. BigMap).

In detail, a library that takes the from of an actor class:

* is separately compiled to wasm, 
* has code included in the client as  a binary blob
* is imported as an asynchronous local function that installs the blob using the IC ManagementCanister interface.

See test/run-drun/distributor.mo for an example.

Also includes a major refactoring of the pipeline to support recursive compilation and better compilation unit detection, just after parsing and before type-checking.

We may want to revisit the form of an actor class library (should it be a module with type and term definition), but that is easily done later.