Allow fetching candid interface from Motoko canisters (#1897)
This implements the ideas in https://github.com/dfinity-lab/motoko/pull/1510, albeit with a method name that clearly marks this as scaffolding.

TL;DR: You can use the query method `__get_candid_interface_tmp_hack : () -> (text)` to fetch the Candid interface.

In the discussion https://dfinity.atlassian.net/browse/NNS-3 the fact that we
still can’t get the interface of a running canister, and thus have no good way
of exploring the implications on developer workflow, no way to develop UI etc.
is embarrasing for a feature that was clear that we want it like 18 months ago.

Why were we blocked? Initially because we expected some form of system-provided
assets (which we then did not get), and then because of disagreement about the
relation between Candid and the raw IC. Also @hansl writes in NNS-3 that the
system-integrated solution requires coordination between too many parties to be
done “soon”.

But this is a very dissatisfying state of affairs, because even if we can’t
nail down the interface right now, we should at least unblock all use cases
that block on this feature.

So in the interest of agility, I want to do, as a temporary work-around, the
implemnetation that we can have now, namely the Candid-internal solution that
does not require system features. Let’s start _using_ this, and then decide and
implement the “right” approach with less pressure.

I see this as consistent with Dom’s request for more agility; with other
scaffolding and experimentation going on right now (`ic1.call_simple`, work on
`exec` etc.), and in a line with such esteemed and useful hacks like
regex-based asset injection, `ic-router` or encoding subnet ids in canister
ids.