Make print_ptr an exported function (#1923)
This makes it possible to call print_pts from the RTS. Without an
exported print_ptr we need different RTS builds for WASI and IC to be
able to implement printers. With print_ptr we can just us it in the RTS
and the compiler would pick the implementation that works for the target
platform (WASI or IC).