Fix collection of BigNums
This is a bug in pointer arithmetic in BigInt collection function where
we confuse words and bytes.

We have types Words and Bytes to avoid this confusion, but I still
managed to break stuff by unwrapping a Words value with .0 and
subtracting the value from a usize value for an address.

A simple regression test added.