take care of (mutually) recursive types
Note that this still doesn't take the types into account that qualify
the components. I.e. repeating the definition with the same field
labels and other types will confuse the debugger.

Contrary  to the above the situation is especially worrisome with
tuples, since these only carry the vector of type references. So when
we have type-recurring tuples, we will loop. E.g.
``` swift
type RecTup = ?(Int, RecTup);
```
I am trying to come up with a solution for this.