Encoder-side functionality for DWARF DIEs (#2071)
Picks up the DIEs coming in via the instruction stream (`Meta`) and building up the DIE hierarchy. This gets subsequently encoded as the `.debug_info` section and accompanied by the various other administrative sections.

Notes added to explain rationale and methodology. Refactorings tho the whole mechanism are in order, which will redo (mostly) the handling of type DIEs.