resurrect funds doc as cycles doc (#2145)
Document Cycle Management (Take II). Let's hope this one sticks to the wall.

I've included the code samples in comments inline for easier reviewing, or use VSCode on the adoc file to have the examples inlines. Choose your poison. I've already drunk mine.

If we get this reviewed soon, we could make another motoko release and make the next train (or miss it in case the Cycles API changes again).