Disallow free-standing blocks (#2102)
BREAKING CHANGE: Blocks are only allowed as sub-expressions of control flow expressions like `if`, `loop`, `case`, etc. In all other places, braces are always considered to start an object literal. To use blocks in other positions, the new `do {<block>}` expression can be used. The more liberal syntax is still allowed for now but deprecated, i.e., produces a warning.

Instead of disambiguating blocks and objects by changing object syntax, change and restrict block syntax: blocks are only allowed as "nested expressions", i.e., as bodies of another control expression. Concretely:

* All control expressions take blocks as their subexpressions.
* As a short-hand, these blocks may have their braces omitted if they are a single expression.
* However, such a short-hand _must not start with an object expression_.
* Free-standing blocks are no longer allowed, they have to be preceded by the `do` keyword.

Consequently, all uses of blocks and objects are unambiguous. This PR warns about the ones that wouldn't be, and marks them as deprecated.

Unparenthesised object patterns remain disallowed in a `func` or `class` declaration (to avoid ambiguity with the following block). For consistency, they are also disallowed after `shared` now.

Examples of allowed syntax:

* `if c { return }`
* `if c { {x = 0} }` (1a)
* `if c return`
* `let x = 1; do { let x = 2 }` (2a)
* `func f({x}) {}`
* `shared({x}) func f() {}` (3a)

Deprecated (will all warn):

* `if c {x = 0}` (1b)
* `if c {x = 0}.x`
* `let x = 1; { let x = 2 }` (2b)
* `shared {x} func f() {}` (3b)

Since there is hardly ever a need to write free-standing blocks in Motoko as (2b), the `do` syntax will not often be needed in practice. (Our test suite is a pathological counter example, because we use such blocks a lot for failure recovery in the fail tests.)

The main impact on practical code hence is the need to change (1b) to (1a) and from (3b) to (3a). Much less intrusive than the changes to object syntax I previously suggested.