Make record syntax non-recursive (#1876)
Changes the record short-hand for objects so that the labels no longer act as identifiers bound in the local scope.

BREAKING CHANGE: Simple object literals of the form `{a = foo(); b = bar()}` do no longer bind the field names locally. This enables writing expressions like `func foo(a : Nat) { return {x = x} }`. OTOH, it breaks expressions like `{a = 1; b = a + 1}`. These now have to be written differently, e.g., with an auxiliary declaration: `let a = 1; {a = a; b = a + 1}`, or using the `object` form: `object {public let a = 1; public let b = a + 1}`.