Auto-update: update `common` less often
we don’t really use the `common` repo, with the exception of making our
OSX builds stand-alone (it’s a in-house tool that wraps the binaries
and the shared libraries in a self-extracting program).

`common` changes a lot, but that code changes rarely. So let’s reduce
repo noise by bumping only once per week.

(I briefly tried to get rid of the `common` dependency alltogether in
https://github.com/dfinity/motoko/pull/2171, but that got stuck
because I could not find out how to do static linking of
non-system-libraries on OSXa)