Move closure table to Rust
This also changes how we store the next free index a little bit. The
constant `FULL` is gone, when the next free location is equal to the
array length that's how we know the table is full now. Also, `FREE_SLOT`
no longer next free location shifted left, it holds the next free
location directly (not shifted).