Use MutBox index offset in obj_idx_ind (#2221)
As described in the module documentation, mutable fields in objects
point to `ObjInd`s or `MutBox`s (which have the same layout), so when
indexing a mutable field we follow the pointer in the field and skip the
`ObjInd` or `MutBox` header. Previously the assumption was the header is
one word (or the field of `MutBox` and `ObjInd` is at offset 1). The
assumption is now removed and we use `MutBox.field` to index the field.