Nix: Stop using standaloneRust, not needed (#2171)
previously , we used the `common` repo to ge the `standaloneRust` tool. That takes a Darwin program, finds all libraries it uses (excluding System libraries), and creates a self-extracting binary from that.

Turns out that our Ocaml binaries have _no_ library dependencies besides system libraries! So we don't need that.

The binaries still refer to some nix paths, but these references are not used, it seems. We can use `remove-references-to` to invalidate them, and use `allowedRequisites` to assert that the binaries are _really_ self-contained.

Dropping the dependency on `common` will reduce repo noise and is a good step towards open sourcing this.